requirements.txt
numpy
pdf2jpg
Pillow
streamlit
st-pages
streamlit_code_editor
streamlit-toggle-switch
